Fix bug in pixelwise image comparison
I was caching sets of points from each candidate image. However, each
piece of the ORIGINAL image I'd process, I was picking a DIFFERENT set
of points. This meant I was comparing a random set of points from the
original piece with a completely different set of points in each
candidate image. This defeats the purpose of doing pixelwise
comparisons.

Fixed it so that I'm using the SAME set of (x, y) ordered pairs
throughout the construction of the mosaic (when using pixelwise
comparison). This ensures that the points we're testing in the original
piece match up with the points we're testing in the candidate images, so
that the pixelwise comparison makes sense.