Add preprocessing of pixelwise data
Can now save "pixelwise" samples of images to a database, and not just
overall "average" colors.

Thus, you can now do preprocessing with the "-w"/pixelwise option.
Samples of pixels will then be saved to the db, so that you don't have
to resample the images when making your next mosaic.