"Preprocessing" always saves pixels and averages
This simplifies things a bit for the user (who now doesn't have to worry
about whether their saved preprocessing files are pixelwise or just have
averages).

It's not a big cost - we have to sample a bunch of pixels in order to
get average colors, so might as well save each pixel to the file so that
we can also use the data for pixelwise comparisons.

Also, starting to make Mosaigraph degrade gracefully if imports fail.