Separate saving to preprocessing file and using it
Now there is an option "-P" that SAVES to a preprocessing file.

"-p", OTOH, now only USES images from the preprocessing file and won't
save any additional images to it. This makes the interface a little
cleaner, I think, since now we don't need a separate option to say DON'T
save to the preprocessing file.