Simplified get_best_edge_length
Instead of trying to calculate an edge length algebraically at first,
and then trying to zero in on the best value with trial and error, just
start and 1 and trial-and-error up to the best value.

This is much simpler, since we know we're always going up.

Note, this new get_best_edge_length sometimes gives different answers
from the old. However, the number of pieces that results should be no
further away from the target number of piece than it was before this
commit.